nit: move getPackageJson to configHelpers
